+++ includes/utils.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nfukada <nfukada@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*   By: totaisei <totaisei@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/12 20:13:16 by nfukada           #+#    #+#             */
-/*   Updated: 2021/02/14 01:19:50 by nfukada          ###   ########.fr       */
+/*   Updated: 2021/02/15 17:24:33 by totaisei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,5 +33,6 @@ void			del_env(t_env **envs, char *name);
 t_env			*get_last_env(t_env *envs);
 size_t			get_env_size(t_env *envs);
 t_env			*create_new_env(char *env_str);
+char	*search_env(t_env *envs, char *name);
 
 #endif
